refactor: refactor code and add comments
