+++ src/index.js
@@ -1,7 +1,7 @@
 // This file consolidates and re-exports components from different files so
 // that they are easier to import in App.jsx
-export * from "./pages/Home";
-export * from "./pages/About/About";
+export * from "./pages/home/Home";
+export * from "./pages/about/About";
 export * from "./pages/sdg11/Sdg11";
-export { default as Content } from "./pages/Content";
+export { default as Content } from "./pages/content/Content";
 export { default as Quiz } from "./pages/quiz/Quiz";
