+++ src/pages/home/Home.jsx
@@ -1,16 +1,24 @@
+import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import City from "/src/assets/images/City_scape.png";
-import goal_11 from "/src/assets/goals/Goal_11.svg";
+
+import db from "../../../firebase/firebaseConfig.js";
+import { collection, doc, getDoc } from "firebase/firestore";
+
 import ListOfGoals from "@/components/ListOfGoals";
 import { useAuthContext } from "@/AuthProvider";
+import City from "/src/assets/images/City_scape.png";
+import goal_11 from "/src/assets/goals/Goal_11.svg";
+
 import { Button } from "@/components/ui/button";
 import { ArrowRightIcon } from "@heroicons/react/24/outline";
 import { StarIcon } from "@heroicons/react/20/solid";
 
-import { useState, useEffect } from "react";
-import { collection, doc, getDoc } from "firebase/firestore";
-import db from "../../firebase/firebaseConfig.js";
-
+/**
+ * Home component displays a "Hi" message, a grid of SDG modules,
+ * and learner scores for SDG 11 as stars
+ * 
+ * @returns {JSX.Element} The rendered Home component.
+ */
 export const Home = () => {
   const { user, userData, role } = useAuthContext();
 
@@ -21,8 +29,13 @@ export const Home = () => {
     }
   };
 
-  // Fetch learner's sdg 11 scores and count completed targets
   const [goalsWithScores, setGoalsWithScores] = useState(ListOfGoals);
+
+   /**
+   * Fetches learner's sdg 11 scores from Firestore and updates sdg 11 
+   * with number of completed targets.
+   * @param {string} uid - The unique identifier of the learner.
+   */
   const fetchLearnerScores = async (uid) => {
     try {
       const learnerDocRef = doc(collection(db, "learners"), uid);
@@ -44,12 +57,11 @@ export const Home = () => {
               },
               0
             );
-
             return { ...goal, score: `${completedTargets}/10` };
           }
           return goal;
         });
-        setGoalsWithScores(updatedGoals);
+        setGoalsWithScores(updatedGoals); // Update state with new goals
       } else {
         console.log("No learner data found for this UID.");
       }
@@ -66,6 +78,7 @@ export const Home = () => {
 
   return (
     <div data-testid="home-page">
+
       {/* Renders Hi message based on if their first name or last name exists */}
       <h1 className="pt-4 pb-12">
         {userData?.firstName
@@ -75,6 +88,7 @@ export const Home = () => {
           : "Hi!"}
       </h1>
 
+      {/* SDG 11 Banner */}
       <div className="bg-orange-200 -mx-12 px-4 h-96 flex items-center justify-around">
         <div className="flex flex-col justify-around">
           <h1>
@@ -112,11 +126,10 @@ export const Home = () => {
         </Link>
       </div>
 
+      {/* grid of modules */}
       <h2 className="leading-[1.3] mt-16 mb-8 text-center text-4xl font-bold">
         Your SDGs
       </h2>
-
-      {/* grid of modules */}
       <div className=" grid grid-cols-1 sm:grid-cols-3 lg:grid-cols-5 gap-4">
         {goalsWithScores.map((goal) => (
           <div
