+++ src/pages/quiz/Quiz.jsx
@@ -13,7 +13,7 @@ import {
 } from "firebase/firestore";
 import db from "../../../firebase/firebaseConfig.js";
 import { useAuthContext } from "@/AuthProvider";
-import Question from "./Question.jsx";
+import Question from "./components/Question.jsx";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
