+++ src/pages/sdg11/components/FullScreenDiv.jsx
