+++ src/pages/sdg11/components/GoalInfo.jsx
