+++ src/routes/PrivateRoute.jsx
@@ -1,15 +1,24 @@
+import { Navigate } from "react-router-dom";
 import { useAuthContext } from "@/AuthProvider";
 import LoadingPage from "@/components/LoadingPage";
-import { Navigate } from "react-router-dom";
 
-
+/**
+ * PrivateRoute component that restricts access for pages
+ * based on the user's authentication status and role.
+ * 
+ * @param {Object} props - The component props.
+ * @param {JSX.Element} props.element - The element to render if access is granted.
+ * @param {string} [props.requiredRole] - The role required to access the route (optional).
+ * 
+ * @returns {JSX.Element} The rendered element if access is granted, or a redirection otherwise.
+ */
 const PrivateRoute = ({ element, requiredRole }) => {
   const { user, role, loading } = useAuthContext();
 
   console.log("PrivateRoute - loading:", loading);
   console.log("PrivateRoute - user:", user);
 
-  // loading screen
+  // Show a loading screen while authentication is being checked
   if (loading) {
     return <LoadingPage />;
   }
@@ -19,7 +28,8 @@ const PrivateRoute = ({ element, requiredRole }) => {
     return <Navigate to="/login" />;
   }
 
-  // Check for required role, if provided
+  // If a required role is provided and the user's role does not match, 
+  // redirect to the home page
   if (requiredRole && role !==requiredRole) {
     return <Navigate to="/" />;
   }
